Antarctica CCSM3 case works, now also with the map dist,weight in single precision (for smaller file sizes).

+++ coordinates.f90
@@ -99,7 +99,7 @@ module coordinates
         ! Neighbor arrays (will be allocated to size npts,nmax)
         integer :: nmax 
         integer,  dimension(:,:), allocatable :: i, quadrant, border
-        real(dp), dimension(:,:), allocatable :: dist, weight
+        real(sp), dimension(:,:), allocatable :: dist, weight
 
     end type
